# Reduce memory consumption for populateKeyParts