`populateKeyParts` used join the same path parts over and other to construct nested arrays. Instead, we reuse the last result.

See benchmark comparison:

```
benchmark                old ns/op     new ns/op     delta
BenchmarkLoadFile-16     146821        136975        -6.71%

benchmark                old allocs     new allocs     delta
BenchmarkLoadFile-16     679            675            -0.59%

benchmark                old bytes     new bytes     delta
BenchmarkLoadFile-16     46277         43459         -6.09%
```