@knadh 2021-06-07
Sweet! Released `v1.1.0` with the two performance PRs.